+++ src/oxDNA_analysis_tools/minify.py
@@ -1,47 +1,83 @@
-from oxDNA_analysis_tools.UTILS.readers import ErikReader, cal_confs
 import argparse
 from os import remove, path
+from sys import stderr
+from collections import namedtuple
 from numpy import round
+from oxDNA_analysis_tools.UTILS.oat_multiprocesser import oat_multiprocesser
+from oxDNA_analysis_tools.UTILS.RyeReader import describe, conf_to_str
+from oxDNA_analysis_tools.UTILS.get_confs import get_confs
 
 import time
 start_time = time.time()
 
+ComputeContext = namedtuple("ComputeContext",["traj_info",
+                                              "top_info",
+                                              "d",
+                                              "a"])
+
+def compute(ctx:ComputeContext, chunk_size:int, chunk_id:int):
+    confs = get_confs(ctx.traj_info.idxs, ctx.traj_info.path, chunk_id*chunk_size, chunk_size, ctx.top_info.nbases)
+    
+    for conf in confs:
+        if ctx.d is not None: #round positions
+            conf.positions = round(conf.positions, ctx.d)
+            conf.a1s = round(conf.a1s, ctx.d)
+            conf.a3s = round(conf.a3s, ctx.d)
+        if ctx.a: #discard a vectors
+            conf.a1s -= conf.a1s
+            conf.a3s -= conf.a3s
+
+    out = ''.join([conf_to_str(c) for c in confs])
+    return out
 
 def main():
     parser = argparse.ArgumentParser(prog = path.basename(__file__), description="Compress given configuration.")
     parser.add_argument('trajectory', type=str, nargs=1, help='the trajectory file you wish to analyze')
     parser.add_argument('outfile',    type=str, nargs=1, help='minified file')
-
+    parser.add_argument('-p', metavar='num_cpus', nargs=1, type=int, dest='parallel', help="(optional) How many cores to use")
     parser.add_argument('-a', action = 'store_true', help='Discard a vectors.')
-    parser.add_argument('-p', type=int, nargs=1,  help='Round positions and orientations to the specified number of digits.')
+    parser.add_argument('-d', type=int, nargs=1,  help='Round positions and orientations to the specified number of digits.')
 
     args = parser.parse_args()
 
     traj_file = args.trajectory[0]
     out = args.outfile[0]
-    # get the number of configurations
-    n_confs = cal_confs(traj_file)
 
-    try: # make sure there is no out file
+    top_info, traj_info = describe(None, traj_file)
+
+    # -p sets the number of parallel processes
+    if args.parallel:
+        ncpus = args.parallel[0]
+    else:
+        ncpus = 1
+
+    # -d sets the decimals of the output
+    if args.d:
+        d = args.d[0]
+    else:
+        d = None
+
+    # -a sets the a vectors to 0
+    if args.a:
+        a = True
+    else:
+        a = False
+
+    try:
         remove(out)
     except:
         pass
-    
-    with ErikReader(traj_file) as reader:
-        for i in range(n_confs):
-            print(i+1,":",n_confs)
-            # Erik reader ignores velocities 
-            system = reader.read()
-            if args.p: # round positions
-                system.positions = round(system.positions, args.p[0])
-                system.a1s = round(system.a1s, args.p[0])
-                system.a3s = round(system.a3s, args.p[0])
-            if args.a: # discard a vectors
-                system.a1s -= system.a1s 
-                system.a3s -= system.a3s
-            # output conf
-            system.write_append(out)
 
+    ctx = ComputeContext(traj_info, top_info, d, a)
+
+    with open(out, 'w+') as f:
+        def callback(i, r):
+            nonlocal f
+            f.write(r)
+
+        oat_multiprocesser(traj_info.nconfs, ncpus, compute, callback, ctx)
+
+    print(f"INFO: Wrote aligned trajectory to {out}", file=stderr)
     print("--- %s seconds ---" % (time.time() - start_time))
 
 if __name__ == '__main__':
